backend/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
